Implement the base game mechanic in javascript
A Game constructor has been defined, to create a new game with a
random distribution of the items in the containers.
Note that the game is not guaranteed to be solvable.

The created object contains a 'table' array, whose elements are the
containers represented by arrays of items.
The items are just numbers each of which is associated with a specific
color.

Methods to implement the next functionalities have been implemented:
- make a move between two containers. This also checks if the move is
  feasible.
- make the reverse move. Again, this also checks if the move is feasible.
- check if the current 'table' is solved

No graphical UI has been created yet.