Setup the graphic interface of the containers and items
Setup a dummy html structure for the containers and the items.
This is just a temporary structure, used to setup the style.
The actual one will be created dynamically via Javascript
depending on the game parameters and setup.

The style of these elements has therefore been implemented.
The rest of the user interface is not implemented yet.

Also, note that the page is not responsive yet. This is mainly due
to the items having a fixed width, and the remaining of dimensions
depending on this value.