Add function to initialize the items' color via javascript
Add a function to initialize the classes .C0, .C1, ... representing
each color an item can have, via Javascript.

A predefined list of 20 distinct colors have been hard-coded in the
script.js file. Also, the code creates a new style element in the
<head>, defining the specific rules.

The function is called when initializing the page.

This way it is easier to change the color palette, if needed.

The corresponding CSS rules in styles.css, e.g.,

	.C0{
	    background-color: var(--col-accent-dark);
	}

have been removed, as they are not needed anymore.