Add the JS code to generate the table with the containers and items
The code to generate the HTML code representing the table with the
containers and items has been implemented.
The structure is generated based on the game.table table in the
Game object.

The HTML code in index.html describing the table has been removed,
as it is no longer necessary, since it is now genrated dynamically.

A function to cancel the current game (both the Game object and the
user interface) has been coded, too. This is necessary when the user
will start a new game.