Add functionality to play the game
An event listener for the click event has been added to each
container.

The callback either sets the 'from' container from a move or determines
whether a move can be performed and applies it if it can be done.

Some delays have been applied through the function
	setTimeout(..., delay)
to ensure some animations are displayed in order.