Add the computation of the items (balls) size to fit the space
The size of the items (--item-len) was fixed. Now it is computed
via CSS so that the table of the game fit the whole page.

It is assumed that containers are organized in --nrows rows and
--ncols = ceil(--C/--nrows) columns, where --C is the number of
containers.

Also, most of the properties of the containers/items are defined
as a fraction of the items length (--item-len), also to simplify
the following computations. These are defined by the variable
names ending with ...-frac.

First, the size that the items should have to fill the available
horizontal space is computed (--item-len-horfill). The available
horizontal space is equal to the --max-width of the page content
minus the table left/right padding and the sum of the border of
the containers in a row (in general, the fixed width dimensions
are to be subtracted here). This available space is divided by
the total occupied space of the remaining items, normalized by
--item-len (ie, considering the ...-frac variables). This way
we find --item-len-horfill.

Second, the size that the items should have to fill the available
vertical space is computed (--item-len-verfill). This is done
similarly to the previous case. However, the available space is
the total viewport height, minus the header, nav and footer height,
as well as the table padding.

The minimum of the two (--item-len-horfill, --item-len-verfill)
is used as --item-len, in order  to fit in both the horizontal and
vertical available space.

NOTE: the following details are to be implemented next:

1) impose the maximum number of columns per row. Right now, the
items are added in the first row as long as there is space, and
might exceed --ncols (ie, because the smaller --item-len-verfill
is used as --item-len instead of --item-len-horfill). This might
result in a non-equal distribution of the items among the rows.

2) (possibly) Set --nrows equal to 1 or 2 depending on the aspect
ratio. Currently this is hardcoded to 2 on css. However, this
might result in a waste of space in ultra-wide aspect ratios.

3) set the values of --C and --ncols via JS (they are currently
hardcoded on the css file). --C should be set when starting a new
game. --nrows should be the rounded up value of --C/--nrows.
Despite the round() function exist in css, it is not widely
supported yet.