Add functionalities to make the page responsive
Three main additions have been implemented.

1) the exact number of columns per row of containers is imposed.
This avoids non-equal distribution of the items among the rows.
To do this, a grid is imposed for the containers of the .table
element, instead of a flexbox, and the number of columns is set
using the repeat(ncols, 1fr) function.

2) select the size of items (balls) and the number of rows of
containers in order to maximize the used space. See function
setNumberOfRowsAndCols() in the JS file js/script.js for more
details.

3) set the computation of the value of --ncols via JS, given the
number of rows and containers. This requires taking the ceiling
of C/nrows. Despite the round('up',...) function exist in css, it
is not widely supported yet, so JS is used instead.