Add the functionality to move more than one item at a time
The global variable SingleMove is introduced.

If it is 'true', only one item at a time can move in a single move.
This is the behaviour implemented so far.

If it is 'false', when there are multiple items in the from container
that can be moved to the destination container (e.g., they have the
same color and there is enough space in the destination), all of
them are moved.

Note that the 'undo last move' functionality is not adapted yet: it
still undoes only the last move. This is going to be fixed next.