Add a dialog box to set the settings of the game
The dialog to select:
- the number of colors,
- the number of empty container,
- the maximum number of items per container,
- the mode in which as many items as possible are moved,
- the mode in which initially only the top items are shown.

Only the style has been setup, and the open/close dialog functionality.

To sliders have been implemented using the noUI-slider library, which
helps ensure support for touch devices.
Instead, a simple custom style has been setup for the checkboxes.

Note that the functionality of the 'save' button in the dialog is
not implemented yet, which actually sets the values set by the user.