Bugfix in the 'moveAsManyItemsAsPossible' and 'hideInitial' mode
The following bug has been resolved:

BUG: When both 'moveAsManyItemsAsPossible' and 'hideInitial' are
true, it might happen that multiple items can be moved between
containers, even if some of them are hidden. Actually, only visible
items could be moved.

Before this commit, the visibility status of the items was handled
only on the style of the items.

Now, it is handled also on the game logic. A counter has been
included for each container, counting the number of hidden items
in there. Note that these are always at the bottom of the container.

Then a move is allowed only if the number of items in the container
is greater than the counter of hidden items in there.

That counter of hidden items is also updated as follows: if the
number of items after the move is greater than zero and equal to
the counter of hidden items, decrease such counter by 1.