Add the dialog showing when no useful move is possible
A useful move is a move that leaves no item of the same color in
the from container. If no move of this kind are possible, there
can be two possibilities:

1) no move is possible (the user was already warned by this)

2) the only possible moves move the items back and forth between the
containers. These are useless move and occur when they leaves at
least one item of the same color in the from container, because
there is not enough space in the to container.

Two arrays (equalItemsOnTop, emptyItemsOnTop) and a 2d matrix
(initialTypeOfMove) have been added to compute whether a move is not
possible, or possible and useful or not, ie, by counting the items
of the same color that remain in a container after it.

A different dialog is shown depending on whether no more moves are
possible, or either the only remaining moves are useless.