Add the functionality to add empty containers to help solving
For some combinations of N (num. of colors), K (num. of empty
containers),H (capacity of containers) some instances of the game
could not be solvable.

The minimum number of empty containers K* to have to be able to
solve EVERY istance of the game having fixed H and N is given by
ceil((H-1)/H*N)

|Source: Ito, T et al, "Sorting Balls and Water: Equivalence and
|Computational Complexity",arXiv:2202.09495 [cs.CC], available
|at https://arxiv.org/pdf/2202.09495.pdf

As the instances of the game are generated randomly, this
possibility could occur. Then, we let the user add empty containers
to help solving the game.

The number of containers that can be added is upper bounded by K*.