Minor fix
The number of empty containers and containers was not properly
re-initialized when new empty containers were added, and the game
was restarted or a new game was started.

This also affected the visualization of the game.

This has been fixed by not modifying K, but including a property
KA of the Game objects counting the containers added, and properly
adjusting the code.